# packages/contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "RouteTarget": {
        "additionalProperties": false,
        "properties": {
          "priority": {
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Priority",
            "type": "integer"
          },
          "target_id": {
            "format": "uuid",
            "title": "Target Id",
            "type": "string"
          }
        },
        "required": [
          "target_id",
          "priority"
        ],
        "title": "RouteTarget",
        "type": "object"
      },
...
      "SetRoute": {
        "additionalProperties": false,
        "properties": {
          "revision": {
            "maximum": 9007199254740991.0,
            "minimum": 1.0,
            "title": "Revision",
            "type": "integer"
          },
          "targets": {
            "items": {
              "$ref": "#/components/schemas/RouteTarget"
            },
            "maxItems": 8,
            "title": "Targets",
            "type": "array"
          }
        },
        "required": [
          "revision",
          "targets"
        ],
        "title": "SetRoute",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "hearth control plane",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/capability-routes/{capability_id}": {
      "put": {
        "operationId": "assign_api_v1_capability_routes__capability_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "capability_id",
            "required": true,
            "schema": {
              "title": "Capability Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetRoute"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Assign",
        "tags": [
          "providers"
        ]
      }
    },
...
  }
}